requirements.txt
streamlit
streamlit-audiorecorder
google-generativeai
pydub